# api/openAPISpec.yml
openapi: 3.0.4
info: 
  title: Auth-service sut lk
  version: 0.0.1
servers:
  - url: http://localhost:{port}/
    description: Local host for testing
    variables:
      port:
        default: "8080"
  - url: http://localhost:{port}/api/{version}/auth
    description: Default local host for testing
    variables:
      port:
        default: "80"
        enum:
          - "80"
          - "8080"
      version: 
        default: 0.0.1
paths:
  /ping:
    get:
      responses:
        "200":
          description: Checking server availability
          content:
            text/plain: 
              schema:
                type: string
                example: Pong
  /login:
    post:
      summary: Login user by email and password
      requestBody:
        required: true
        content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  password:
                    type: string
                example:
                  email: "example@example.com"
                  password: "password"
      responses:
        "200":
          description: Return Access and Refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokens"
        "401":
          description: Email or password are incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/err_msg"
  /refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              example: 
                refresh_token: <refresh jwt token>
      responses:
        "200":
          description: Return Access and Refresh tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/tokens"
        "401":
          description: Bad refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/err_msg"
  /logout:
    post:
      summary: Will brake all created access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                access_token:
                  type: string
              example:
                  access_token: <access_token>
      responses:
        "200":
          description: All tokens had been expired
        "400":
          description: Something went wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/err_msg"
  /checktoken:
    post:
      summary: Validating JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:  
                type: object
                properties:
                  token:
                    type: string
      responses:
        "200":
          description: Token is valid
        "400":
          description: Token is invalid

components:
  schemas:
    tokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
    err_msg:
      type: object
      properties:
        code: 
          type: integer
        msg:  
          type: string